llm_webservice/requirements.txt
# Core ML & embeddings
accelerate
transformers
sentence-transformers
chromadb

# Web framework
flask
waitress

# Other utilities
python-docx

